+++ app/src/main/java/com/zeapo/pwdstore/PasswordFragment.kt
@@ -5,6 +5,7 @@
 package com.zeapo.pwdstore
 
 import android.content.Context
+import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -16,7 +17,10 @@ import androidx.preference.PreferenceManager
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
+import com.zeapo.pwdstore.git.GitActivity
 import com.zeapo.pwdstore.ui.adapters.PasswordRecyclerAdapter
 import com.zeapo.pwdstore.utils.PasswordItem
 import com.zeapo.pwdstore.utils.PasswordRepository
@@ -44,6 +48,7 @@ class PasswordFragment : Fragment() {
     private lateinit var recyclerView: RecyclerView
     private lateinit var listener: OnFragmentInteractionListener
     private lateinit var settings: SharedPreferences
+    private lateinit var swipe_refresher: SwipeRefreshLayout
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -61,6 +66,19 @@ class PasswordFragment : Fragment() {
         val view = inflater.inflate(R.layout.password_recycler_view, container, false)
         // use a linear layout manager
         val layoutManager = LinearLayoutManager(requireContext())
+        swipe_refresher = view.findViewById(R.id.swipe_refresher)
+        swipe_refresher.setOnRefreshListener {
+            if (!PasswordRepository.isInitialized) {
+                MaterialAlertDialogBuilder(requireContext())
+                        .setMessage(getString(R.string.creation_dialog_text))
+                        .setPositiveButton(getString(R.string.dialog_ok), null).show()
+                swipe_refresher.isRefreshing = false
+            } else {
+                val intent = Intent(context, GitActivity::class.java)
+                intent.putExtra("Operation", GitActivity.REQUEST_SYNC)
+                startActivityForResult(intent, GitActivity.REQUEST_SYNC)
+            }
+        }
         recyclerView = view.findViewById(R.id.pass_recycler)
         recyclerView.layoutManager = layoutManager
         // use divider
@@ -124,6 +142,10 @@ class PasswordFragment : Fragment() {
         } catch (e: ClassCastException) {
             throw ClassCastException("$context must implement OnFragmentInteractionListener")
         }
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        swipe_refresher.isRefreshing = false
     }
 
     /** clears the adapter content and sets it back to the root view  */