+++ app/src/main/java/com/zeapo/pwdstore/crypto/PgpActivity.kt
@@ -40,6 +40,7 @@ import com.zeapo.pwdstore.PasswordEntry
 import com.zeapo.pwdstore.R
 import com.zeapo.pwdstore.UserPreference
 import com.zeapo.pwdstore.ui.dialogs.PasswordGeneratorDialogFragment
+import com.zeapo.pwdstore.ui.dialogs.XkPasswordGeneratorDialogFragment
 import com.zeapo.pwdstore.utils.Otp
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
@@ -150,8 +151,12 @@ class PgpActivity : AppCompatActivity(), OpenPgpServiceConnection.OnBound {
                 setContentView(R.layout.encrypt_layout)
 
                 generate_password?.setOnClickListener {
-                    PasswordGeneratorDialogFragment()
-                      .show(supportFragmentManager, "generator")
+                    when (settings.getString("pref_key_pwgen_type", KEY_PWGEN_TYPE_CLASSIC)) {
+                        KEY_PWGEN_TYPE_CLASSIC -> PasswordGeneratorDialogFragment()
+                                .show(supportFragmentManager, "generator")
+                        KEY_PWGEN_TYPE_XKPASSWD -> XkPasswordGeneratorDialogFragment()
+                                .show(supportFragmentManager, "xkpwgenerator")
+                    }
                 }
 
                 title = getString(R.string.new_password_title)
@@ -524,8 +529,12 @@ class PgpActivity : AppCompatActivity(), OpenPgpServiceConnection.OnBound {
     private fun editPassword() {
         setContentView(R.layout.encrypt_layout)
         generate_password?.setOnClickListener {
-            PasswordGeneratorDialogFragment()
-              .show(supportFragmentManager, "generator")
+            when (settings.getString("pref_key_pwgen_type", KEY_PWGEN_TYPE_CLASSIC)) {
+                KEY_PWGEN_TYPE_CLASSIC -> PasswordGeneratorDialogFragment()
+                        .show(supportFragmentManager, "generator")
+                KEY_PWGEN_TYPE_XKPASSWD -> XkPasswordGeneratorDialogFragment()
+                        .show(supportFragmentManager, "xkpwgenerator")
+            }
         }
 
         title = getString(R.string.edit_password_title)
@@ -843,6 +852,9 @@ class PgpActivity : AppCompatActivity(), OpenPgpServiceConnection.OnBound {
         private const val ACTION_START = "ACTION_START_CLIPBOARD_TIMER"
 
         const val TAG = "PgpActivity"
+
+        const val KEY_PWGEN_TYPE_CLASSIC = "classic"
+        const val KEY_PWGEN_TYPE_XKPASSWD = "xkpasswd"
 
         private var delayTask: DelayShow? = null