+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Copy implicit username (password filename) by long pressing
+- Create xkpasswd style passwords
+- Swipe on password list to synchronize repository
 
 ### Fixed
 - Resolve memory leaks on password decryption